+++ src/contents/settings/glossary.tsx
@@ -6,6 +6,7 @@ import { BookOpen } from "lucide-react";
 import { useGlossary } from "@/context/glossary";
 import { lazy, Suspense } from "react";
 import { Skeleton } from "@/components/ui/skeleton";
+import { Switch } from "@/components/ui/switch";
 
 const SelectorField = lazy(()=>import("@/components/fields/selector"))
 
@@ -38,6 +39,17 @@ export default function GlossarySettings(){
                                    ]}
                               />
                          </Suspense>
+                    </SettingsOption>
+                    <SettingsOption
+                         title="Show sidebar by default"
+                         description="Show the glossary sidebar everytime you enter the app"
+                         id="show-glossary"
+                    >
+                         <Switch
+                              id="show-glossary"
+                              checked={settings.showGlossary}
+                              onCheckedChange={checked=>setSettings({showGlossary: checked})}     
+                         />
                     </SettingsOption>
                </SettingsItem>
                {/* TODO: Auto-load glossary */}